apps/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
supabase
transformers
torch
numpy
python-dotenv
